chore[RabbitMQ]: move classes to package for better organization
